+++ internal/local/local_test.go
@@ -1,4 +1,4 @@
-package backend
+package local
 
 import (
 	"testing"
